requirements.txt
crewai[litellm]>=1.10.0
crewai-tools>=0.38.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
streamlit>=1.43.0
requests>=2.32.0
pydantic>=2.10.0
